fix: Monitoring Token 페이지 Cluster name 입력란 디자인 수정
